docs, refactor
