+++ src/page/view/BubbleSort/indexing/types.ts
@@ -2,3 +2,4 @@ export type Receipt<S, C> = {
   initialState: S;
   commands: C[];
 };
+export type CommandExecutor<S, C> = (state: S, command: C) => S;
